changed the type of class constants from double to int so that they can be initialized as part of the class definition.  we should rely on implicit type conversion in this case so this change should be safe.
quite trivial so committing it to trunk directly, but I'll note it on the ML.


git-svn-id: svn://bind10.isc.org/svn/bind10/trunk@2672 e5f2f494-b856-4b98-b285-d166d9295462

+++ src/lib/bench/tests/benchmark_unittest.cc
@@ -60,7 +60,7 @@ BenchMark<TestBenchMark>::tearDown() {
 // XXX: some compilers cannot find class static constants used in
 // EXPECT_xxx macross, for which we need an explicit definition.
 template <typename T>
-const double BenchMark<T>::TIME_FAILURE;
+const int BenchMark<T>::TIME_FAILURE;
 }
 }
